requirements.txt
flask
flask-restful
flask-security-too
flask-sqlalchemy
flask-jwt-extended
bcrypt